Feat: Merged , Smart Part Payment and Transactional generater into one
